requirements.txt
Django==4.1.9
mysqlclient==2.1.1
dj-database-url==1.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.7
asgiref==3.6.0
sqlparse==0.2.4
tzdata==2022.7
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.59.1
matplotlib-inline==0.1.6
numpy==1.25.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.76
pandas==2.0.3
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
regex==2023.6.3
requests==2.31.0
